requirements.txt
pandas
fastapi
uvicorn
sqlalchemy
matplotlib
jinja2
httpx
python-dotenv
google-generativeai